+++ priv/repo/migrations/20210409074413_add_unique_index_to_shared_link_name.exs
@@ -1,0 +1,7 @@
+defmodule Plausible.Repo.Migrations.AddUniqueIndexToSharedLinkName do
+  use Ecto.Migration
+
+  def change do
+    create unique_index(:shared_links, [:site_id, :name], name: :shared_links_site_id_name_index)
+  end
+end